+++ data/icons.php
@@ -1,5 +1,5 @@
 <?php
-/* Last Updated : 03/03/2021 - 11:07:20:am */
+/* Last Updated : 08/03/2021 - 06:11:34:pm */
 
 defined( 'ABSPATH' ) || exit;