+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.5.3.6] - 08/03/2021
+## Added
+* Option to bulk add fields in nested fields
+
+## Fixed
+* Color Picker Field (many issues with its possible options)
+* WC Product Metabox Ajax Loading Issue
+
 # [1.5.3.5] - 28/02/2021
 ## Fixed
 * Fixed Dashboard Widget Loading if no fields are set